Finish the cloud code refactor
