+++ zephyr/__init__.py
@@ -1,3 +1,3 @@
-from zephyr.scale import API_V1, API_V2, ZephyrScale
-from zephyr.squad import ZephyrSquad
+from zephyr.scale import API_V1 as SCALE_API_V1, API_V2 as SCALE_API_V2, ZephyrScale
+from zephyr.squad import API_V1 as SQUAD_API_V1, API_V2 as SQUAD_API_V2, ZephyrSquad
 from zephyr.utils.common import cookie_str_to_dict
